Refactor into ui package
